Move status bar into module
