+++ src/platform/mod.rs
@@ -11,6 +11,7 @@
 //! And the following platform-specific module:
 //!
 //! - `run_return` (available on `windows`, `unix`, `macos`, and `android`)
+//! - `status_bar` (available on `windows`, `unix`, and `macos`)
 //!
 //! However only the module corresponding to the platform you're compiling to will be available.
 
@@ -21,3 +22,4 @@ pub mod unix;
 pub mod windows;
 
 pub mod run_return;
+pub mod status_bar;
