+++ src/platform/unix.rs
@@ -6,10 +6,7 @@
   target_os = "openbsd"
 ))]
 
-use crate::{
-  error::OsError, event_loop::EventLoopWindowTarget, menu::MenuItem, platform_impl, window::Window,
-};
-use std::path::PathBuf;
+use crate::window::Window;
 
 pub use crate::platform_impl::hit_test;
 
@@ -29,38 +26,5 @@ impl WindowExtUnix for Window {
 
   fn skip_taskbar(&self) {
     self.window.skip_taskbar()
-  }
-}
-
-/// Status bar is a system tray icon usually display on top right or bottom right of the screen.
-#[derive(Debug, Clone)]
-#[cfg(target_os = "linux")]
-pub struct Statusbar {
-  pub(crate) icon: PathBuf,
-  pub(crate) items: Vec<MenuItem>,
-}
-
-pub struct StatusbarBuilder {
-  status_bar: Statusbar,
-}
-
-impl StatusbarBuilder {
-  /// Creates a new Statusbar for platforms where this is appropriate.
-  pub fn new(icon: PathBuf, items: Vec<MenuItem>) -> Self {
-    Self {
-      status_bar: Statusbar { icon, items },
-    }
-  }
-
-  /// Builds the status bar.
-  ///
-  /// Possible causes of error include denied permission, incompatible system, and lack of memory.
-  #[inline]
-  pub fn build<T: 'static>(
-    self,
-    _window_target: &EventLoopWindowTarget<T>,
-  ) -> Result<Statusbar, OsError> {
-    platform_impl::Statusbar::initialize(&_window_target.p, &self.status_bar)?;
-    Ok(self.status_bar)
   }
 }
