+++ src/platform_impl/macos/status_bar.rs
@@ -2,7 +2,7 @@ use super::menu::{make_custom_menu_item, make_menu_item, KeyEquivalent};
 use crate::{
   error::OsError,
   menu::{MenuItem, MenuType},
-  platform::macos::Statusbar as RootStatusbar,
+  platform::status_bar::Statusbar as RootStatusbar,
   platform_impl::EventLoopWindowTarget,
 };
 use cocoa::{
